Added error message for install as root

+++ build_install.sh
@@ -7,7 +7,11 @@ if [ $1 == "build" ]; then
 fi
 
 if [ $1 == "install" ]; then
-   echo "root privilege is required for installation"
+   # Make sure only root can run our script
+   if [[ $EUID -ne 0 ]]; then
+      echo "Install script must be run as root" 1>&2
+      exit 1
+   fi
    systemctl stop bserver
    cp bserver /usr/local/bin/
    cp bserver.service /etc/systemd/system/
